refactor functional tests
